refactor(apps/web): Join 페이지 리펙토링 및 이미지 변경
