+++ packages/theme/src/tokens/typography.ts
@@ -7,6 +7,7 @@ export const typography = {
     22: '2.2rem',
     24: '2.4rem',
     28: '2.8rem',
+    44: '4.4rem',
   },
   fontWeight: {
     medium: '500',
