+++ packages/ui/src/components/Text/Text.css.ts
@@ -9,4 +9,5 @@ export const textStyle = style({
   fontWeight: weightVar,
   color: colorVar,
   lineHeight: '150%', // TODO 추후 분리
+  whiteSpace: 'pre-line',
 });
